Refactor

+++ src/main.ts
@@ -3,7 +3,7 @@ import { run } from './run'
 
 const main = async (): Promise<void> => {
   await run({
-    expirationDays: Number.parseInt(core.getInput('expiration-days', { required: true })),
+    expirationDays: Number.parseFloat(core.getInput('expiration-days', { required: true })),
     refPrefix: core.getInput('ref-prefix', { required: true }),
     dryRun: core.getBooleanInput('dry-run'),
     token: core.getInput('token', { required: true }),
